requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
config==0.5.1
Django==4.1.2
django-bootstrap3==22.1
gunicorn==20.1.0
idna==3.4
python-decouple==3.6
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.5
urllib3==1.26.13
